# docs/swagger/src/main/java/it/csi/stacore/staconsultazionesrv/api/dto/Veicolo.java
package it.csi.stacore.staconsultazionesrv.api.dto;

import java.util.Objects;
import java.util.ArrayList;
import io.swagger.annotations.ApiModel;
import java.io.Serializable;
import javax.validation.constraints.*;
import io.swagger.annotations.*;

@ApiModel(description="Veicolo")

public class Veicolo  implements Serializable {
  // verra' utilizzata la seguente strategia serializzazione degli attributi: [implicit-camel-case] 
  private static final long serialVersionUID = 1L;
  
  private String dummy = null;

  /**
   * Dummy prop
   **/
  
  @ApiModelProperty(value = "Dummy prop")

  // nome originario nello yaml: dummy 
  public String getDummy() {
    return dummy;
  }
  public void setDummy(String dummy) {
    this.dummy = dummy;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Veicolo veicolo = (Veicolo) o;
    return Objects.equals(dummy, veicolo.dummy);
  }

  @Override
  public int hashCode() {
    return Objects.hash(dummy);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Veicolo {\n");
    
    sb.append("    dummy: ").append(toIndentedString(dummy)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}
