# docs/swagger/staconsultazionesrv-1.0.0-swagger.yaml
swagger: '2.0'
info:
  description: API componente staconsultazionesrv
  version: 1.0.0
  title: STACONSULTAZIONESRV
  contact:
    email: andrea.fontanelli@eng.it
# tags are used for organizing operations
tags:
- name: test
  description: test resources
- name: posizioni
  description: gestione posizioni
  

# basePath: /api
# schemes:
#    - http

paths:
  /test/resources:
    get:
      tags:
        - test
      summary: Test resources
      description: Effettua il test dei servizi richiamati
      operationId: testResources
      produces:
      - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/GenericResponse'
          description: OK
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'
  /posizione/{id}:
    get:
      tags:
        - posizioni
      summary: Posizione 
      description: Effettua il test dei servizi richiamati
      operationId: caricaDettaglioPosizione
      consumes:
        - application/json
      produces:
      - application/json
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: id posizione da caricare
          format: int64
        - in: body
          name: utente
          description: Utente
          required: true
          schema:
            $ref: '#/definitions/Utente'
          
      responses:
        200:
          schema:
            $ref: '#/definitions/Posizione'
          description: OK
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'
definitions:
  Posizione:
    type: object
    description: Posizione
    properties:
      id:
        type: integer
        format: int64
        example: id Posizione
      idOperazioneLogica:
        type: integer
        format: int64
        example: id Operazione Logica
      idVeicolo:
        type: integer
        format: int64
        example: id veicolo
      idTipoSoggettoPassivo:
        type: integer
        format: int64
        example: id tipo soggetto passivo
      proprietario:
        $ref: '#/definitions/Proprietario'
        description: Proprietario
      veicolo:
        $ref: '#/definitions/Veicolo'
      fonte:
        $ref: '#/definitions/Fonte'
      eventoInizio:
        $ref: '#/definitions/Evento'
        description: Evento inizio
      eventoFine:
        $ref: '#/definitions/Evento'
        description: Evento fine
      codiceFiscale:
        type: string
        description: codice fiscale
      tipoUtente:
        $ref: '#/definitions/TipoUtente'
      ente:
        $ref: '#/definitions/Ente'
      profilo:
        $ref: '#/definitions/Profilo'
      abilitazione:
        $ref: '#/definitions/Abilitazione'
      elencoCassa:
        type: array
        description: Elenco casse
        items:
          $ref: '#/definitions/Cassa'
      elencoSportello:
        type: array
        description: Elenco sportelli
        items:
          $ref: '#/definitions/Sportello'
  Evento:
    type: object
    description: Evento
    properties:
      dataEvento:
        type: string
        format: date
        description: data evento
      tipoEvento:
        $ref: '#/definitions/TipoEvento'
  TipoEvento:
    type: object
    description: Evento
    properties:
      id:
        type: integer
        format: int32
        example: id Tipo Evento
      codice:
        type: string
        description: Codice evento
      descrizione:
        type: string
        description: Descrizione evento
  Fonte:
    type: object
    description: Fonte
    properties:
      id:
        type: integer
        format: int32
        example: id Fonte
      codice:
        type: string
        description: Codice fonte
      descrizione:
        type: string
        description: Descrizione fonte
  Utente:
    type: object
    description: Utente
    properties:
      nome:
        type: string
        description: Nome Utente
      cognome:
        type: string
        description: Cognome Utente
      descrizione:
        type: string
        description: descrizione Utente
      codiceFiscale:
        type: string
        description: codice fiscale
      tipoUtente:
        $ref: '#/definitions/TipoUtente'
      ente:
        $ref: '#/definitions/Ente'
      profilo:
        $ref: '#/definitions/Profilo'
      abilitazione:
        $ref: '#/definitions/Abilitazione'
      elencoCassa:
        type: array
        description: Elenco casse
        items:
          $ref: '#/definitions/Cassa'
      elencoSportello:
        type: array
        description: Elenco sportelli
        items:
          $ref: '#/definitions/Sportello'
  TipoUtente:
    type: object
    description: Tipo Utente
    properties:
      codice:
        type: string
        description: Codice Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
      descrizione:
        type: string
        description: Descrizione tipo Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
  Ente:
    type: object
    description: Ente
    properties:
      id:
        type: integer
        format: int32
        example: Id ente
      codice:
        type: string
        description: Codice 
      descrizione:
        type: string
        description: ACI
      email:
        type: string
        description: Email
      indirizzo:
        type: string
        description: Indirizzo
        example: 'Indirizzo Ente'
      civico:
        type: string
        description: Civico
        example: 'Civico Ente'
      cap:
        type: string
        description: CAP
      comune:
        type: string
        description: Comune
      codiceFiscale:
        type: string
        description: Comune
      provincia:
        type: string
        description: Provincia
      sigla:
        type: string
        description: Sigla provincia  
      dataInizioAttivita:
        type: string
        format: date
        description: Data inizio attività    
      abilitazioneRiscossione:
        type: boolean
        description: Indica se l'ente è abilitato alla riscossione 
      elencoSportello:
        type: array
        description: Elenco sportelli
        items:
          $ref: '#/definitions/Sportello'
  Abilitazione:
    type: object
    description: Abilitazione
    properties:
      elencoCasse:
        type: array
        description: Elenco casse
        items:
          $ref: '#/definitions/Cassa'
  Profilo:
    type: object
    description: Profilo Utente
    properties:
      elencoAbilitazione:
        type: array
        description: Elenco casse
        items:
          $ref: '#/definitions/Abilitazione'
  Cassa:
    type: object
    description: Cassa
    properties:
      id:
        type: integer
        format: int32
      codice:
        type: string
        description: Codice 
      descrizione:
        type: string
        description: Descrizione cassa
      denominazione:
        type: string
        description: Denominazione cassa
      idUtente:
        type: string
        description: Id utente
  Sportello:
    type: object
    description: Sportello
    properties:
      id:
        type: integer
        format: int32
        description: id sportello
      codice:
        type: string
        description: Codice sportello
      descrizione:
        type: string
        description: descrizione sportello
      elencoCassa:
        type: array
        description: Elenco casse
        items:
          $ref: '#/definitions/Cassa'
  Veicolo:
    type: object
    description: Veicolo
    properties:
      dummy:
        type: string
        description: Dummy prop
        
  Proprietario:
    type: object
    description: Proprietario
    properties:
      dummy:
        type: string
        description: dummy
  PosizioneExt:
    type: object
    description: Posizione Estesa
    properties:
      dummy:
        type: string
        description: Dummy property
  GenericResponse:
    type: object
    description: Risposta di un servizio generico
    properties:
      message:
        type: string
        description: Contiene il messaggio eventuale del servizio chiamato
        example: 'Servizio ha risposto correttamente'
  ErrorDto:
    type: object
    description: Rappresenta l'errore restituito da servizio
    properties:
      status:
        type: integer
        description: status http restituito
        example: '500, 400, ecc'
      code:
        type: string
        description: code http restituito
        example: 'Internal server error, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
      errorDetails:
        type: array
        description: i dettagli errore
        items:
          $ref: '#/definitions/ErrorDetail'
  ErrorDetail:
    type: object
    description: rappresenta il dettaglio errore restituito in caso di errori di validazione sui campi di input
    properties:
      field:
        type: string
        description: il nome del campo per cui è falita la validazione
        example: 'name, surname'
      error:
        type: string
        description: l'errore di validazione effettivo
        example: 'campo obbligatorio, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
  
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
# basePath: /studiofontanelli/stacomunicazionesrv/1.0.0
schemes:
 - https
 - http
# Added by API Auto Mocking Plugin
# basePath: /andreafontanelli/stacomunicazionesrv/1.0.0
# Added by API Auto Mocking Plugin
basePath: /andreafontanelli/staconsultazionesrv/1.0.0